Removed instance vars from Part

+++ code/softwarecraftsmanship/Main.java
@@ -27,6 +27,6 @@ public class Main {
                 StringWrapperFactory.createStringWrapperWithString("MSI H97 PC MATE - Socket 1150 - ATX"), 
                 DoubleWrapperFactory.createDoubleWrapperWithDouble(89.0)));
         
-        System.out.println((computer.isComplete().booleanValue()) ? "Total price is EUR " + computer.getPrice() : "Your computer is not ready yet...");
+        System.out.println((computer.isComplete().booleanValue()) ? "Total price is EUR " + computer.getPrice().doubleValue() : "Your computer is not ready yet...");
     }
 }
